register refactor
