+++ lib/Features/Auth/register/presentation/widgets/terms_checkbox.dart
@@ -1,0 +1,42 @@
+import 'package:flutter/material.dart';
+import 'package:study_mate/Core/Theme/app_colors.dart';
+import 'package:study_mate/Core/Theme/app_text_styles.dart';
+
+class TermsCheckbox extends StatefulWidget {
+  const TermsCheckbox({super.key});
+
+  @override
+  State<TermsCheckbox> createState() => _TermsCheckboxState();
+}
+
+class _TermsCheckboxState extends State<TermsCheckbox> {
+  bool _agreeToTerms = false;
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        Checkbox(
+          value: _agreeToTerms,
+          activeColor: AppColors.primary,
+          onChanged: (bool? value) {
+            setState(() {
+              _agreeToTerms = value ?? false;
+            });
+          },
+        ),
+        const Text("I agree to the "),
+        GestureDetector(
+          onTap: () {},
+          child: Text(
+            "Terms Of Conditions",
+            style: AppTextStyles.poppins12RegularUnderline(
+              color: AppColors.primary,
+              fontSize: 14,
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
